extract session metadata processing into session_info (WIP)
