+++ sintr_analysis_server_example/lib/session_metrics.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library sintr_worker_lib.session_info;
+library sintr_worker_lib.session_metrics;
 
 import 'dart:convert';
