+++ sintr_analysis_server_example/test_driver.dart
@@ -6,14 +6,9 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io' as io;
 
+import 'package:sintr_worker_lib/completion_metrics.dart';
 import 'package:sintr_worker_lib/instrumentation_lib.dart';
-import 'package:sintr_worker_lib/session_info.dart'
-    show
-        sessionInfoExtractionStart,
-        sessionInfoExtraction,
-        sessionInfoExtractionFinished,
-        sessionInfoReducer,
-        sessionInfoReductionMerge;
+import 'package:sintr_worker_lib/session_info.dart';
 
 main(List<String> args) async {
   if (args.length != 1) {
@@ -26,8 +21,8 @@ main(List<String> args) async {
   var sessionInfo = await loadSessionInfo(path);
   var f = new io.File(path);
 
-  sessionInfoExtractionStart(sessionInfo);
-  LogItemProcessor proc = new LogItemProcessor(sessionInfoExtraction);
+  completionExtractionStart(sessionInfo);
+  LogItemProcessor proc = new LogItemProcessor(completionExtraction);
   var extracted = <String>[];
 
   // Process each line to extract information
@@ -40,7 +35,7 @@ main(List<String> args) async {
   processMessages(proc, extracted);
 
   // Finish the extraction process
-  var finalResults = sessionInfoExtractionFinished();
+  var finalResults = completionExtractionFinished();
   if (finalResults is String) {
     print(finalResults);
   } else if (finalResults is List<String>) {
@@ -55,18 +50,18 @@ main(List<String> args) async {
   var reduced2 = <String, dynamic>{};
   int index = 0;
   while (index < extracted.length / 2) {
-    reduced1 = sessionInfoReducer(extracted[index], reduced1);
+    reduced1 = completionReducer(extracted[index], reduced1);
     ++index;
   }
   print(reduced1);
   while (index < extracted.length) {
-    reduced2 = sessionInfoReducer(extracted[index], reduced2);
+    reduced2 = completionReducer(extracted[index], reduced2);
     ++index;
   }
   print(reduced2);
 
   // Merge the result maps
-  var reduced = sessionInfoReductionMerge(reduced1, reduced2);
+  var reduced = completionReductionMerge(reduced1, reduced2);
   print(reduced);
 }
 
@@ -80,52 +75,7 @@ Future<Map> loadSessionInfo(String path) async {
 
   var priPath = '${path.substring(0, index)}-PRI$sessionId.json';
   var f = new io.File(priPath);
+  Stream<List<int>> stream = f.openRead();
 
-  var firstLine;
-  LogItemProcessor proc = new LogItemProcessor((line) {
-    if (firstLine == null) firstLine = line;
-  });
-  await for (String ln
-      in f.openRead().transform(UTF8.decoder).transform(new LineSplitter())) {
-    proc.addRawLine(ln);
-    processMessages(proc);
-  }
-  proc.close();
-  processMessages(proc);
-
-  var data = firstLine.split(':');
-  return {
-    'sessionId': sessionId,
-    'clientStartTime': data[0].substring(1),
-    'uuid': data[2],
-    'clientId': data[3],
-    'clientVersion': data[4],
-    'serverVersion': data[5],
-    'sdkVersion': data[6]
-  };
-}
-
-void processMessages(LogItemProcessor proc, [List<String> extracted]) {
-  String nextMessage;
-  while (proc.hasMoreMessages) {
-    try {
-      nextMessage = null;
-      nextMessage = proc.readNextMessage();
-    } catch (e, st) {
-      var exMsg = e.toString();
-      if (exMsg.length > 300) exMsg = '${exMsg.substring(0, 300)} ...';
-      print("Error in line \n${exMsg} \n$st");
-    }
-
-    if (nextMessage != null) {
-      print(nextMessage);
-      if (extracted != null) extracted.add(nextMessage);
-    }
-    // if (nextMessage != null) print("${nextMessage[0]}, ${nextMessage[1]}");
-    //
-    // String messageType = nextMessage[1];
-    // msgTyps.putIfAbsent(messageType, () => msgTyps.length);
-    // print("${nextMessage[0]}, ${msgTyps[messageType]}");
-
-  }
+  return await readSessionInfo(sessionId, stream);
 }
